requirements.txt
# Core Flask dependencies
Flask==2.3.2
Werkzeug==2.3.7
Flask-Login==0.6.2

# Environment and configuration
python-dotenv==1.0.0

# Data processing
pandas==2.0.3

# Production server (optional)
waitress==2.1.2

# Development and testing (optional)
# pytest==7.4.0
# pytest-flask==1.2.0
